refactor: UI 관련한 객체를 view 패키지에 구현
