+++ src/main/java/racingcar/step4/RacingCarGameApplication.java
@@ -1,29 +1,12 @@
 package racingcar.step4;
 
-import racingcar.step4.domain.Car;
-import racingcar.step4.domain.RacingCarGame;
-import racingcar.step4.domain.factory.CarFactory;
-import racingcar.step4.domain.factory.RacingCarGameFactory;
-import racingcar.step4.presentaion.InputValue;
-import racingcar.step4.presentaion.InputView;
-import racingcar.step4.presentaion.ResultView;
-
-import java.util.List;
+import racingcar.step4.controller.RacingCarGameController;
+import racingcar.step4.view.InputView;
+import racingcar.step4.view.ResultView;
 
 public class RacingCarGameApplication {
-
   public static void main(String[] args) {
-    new RacingCarGameApplication().run(args);
-  }
-
-  public void run(String[] args) {
-    InputValue input = InputView.receive();
-
-    List<Car> cars = CarFactory.of(input.getInputName());
-    List<RacingCarGame> racingCarGames = RacingCarGameFactory.of(input.getTryCount(), cars);
-    racingCarGames.forEach(RacingCarGame::start);
-
-    ResultView.printResult(racingCarGames);
-    ResultView.printFinalWinners(racingCarGames);
+    RacingCarGameController racingCarGameController = new RacingCarGameController(new InputView(), new ResultView());
+    racingCarGameController.run();
   }
 }
