+++ src/main/java/racingcar/step4/docs/README.md
@@ -3,16 +3,16 @@
 
 ### 기능 요구 사항
 <hr />
-<br />
 
 - 각 자동차에 이름을 부여할 수 있다. **자동차 이름은 5자**를 초과할 수 없다. 
 - 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다. 
 - 자동차 이름은 쉼표(,)를 기준으로 구분한다. 
 - **자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한명 이상일 수 있다.**
 
+<br />
+
 ### 구현 목록
 <hr />
-<br />
 
 - [X] **indent(인덴트, 들여쓰기) depth를 2를 넘지 않도록 구현한다. 1까지만 허용한다.**
   - 예를 들어 while문 안에 if문이 있으면 들여쓰기는 2이다. 
@@ -28,7 +28,18 @@
   - 힌트: if 조건절에서 값을 return하는 방식으로 구현하면 else를 사용하지 않아도 된다.
   - else를 쓰지 말라고 하니 switch/case로 구현하는 경우가 있는데 switch/case도 허용하지 않는다.
 
+<br />
+
+
+### 리팩토링 요구사항
 <hr />
+
+- [X] 핵심 비지니스 로직을 가지는 객체를 domain 패키지에 구현한다.
+- [X] UI 관련한 객체를 view 패키지에 구현한다.
+- [ ] MVC 패턴 기반으로 리팩토링
+  - **domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현하기**
+- [ ] 테스트 가능한 부분과 테스트하기 힘든 부분을 분리
+  - 테스트 가능한 부분에 대해서만 단위 테스트를 진행
 
 ### 새롭게 알게된 내용
 - [X] Object.equals()
